+++ src/Metadata/WikiData/WikiDataCache.php
@@ -24,6 +24,36 @@ class WikiDataCache extends BaseCache
         'wikidata/series/title',
         'wikidata/entities',
     ];
+    /** @var array<string, mixed> */
+    public static array $typeDefinitions = [
+        'author' => [
+            'Q5' => 'human',
+        ],
+        'book' => [
+            'Q7725634' => 'literary work',
+            'Q47461344' => 'written work',
+            // 'Q3331189' => 'version, edition, or translation',
+            // 'Q1279564' => 'short story collection',
+            // 'Q49084' => 'short story',
+        ],
+        'series' => [
+            'Q1667921' => 'novel series',
+            'Q277759' => 'book series',
+            // 'Q13593966' => 'literary trilogy',
+        ],
+        'publisher' => [
+            'Q2085381' => 'publisher',
+        ],
+    ];
+    /** @var array<string, string> */
+    public static array $instanceTypes = [
+        'Q5' => 'author',
+        'Q7725634' => 'book',
+        'Q47461344' => 'book',
+        'Q1667921' => 'series',
+        'Q277759' => 'series',
+        'Q2085381' => 'publisher',
+    ];
 
     /**
      * Summary of getAuthorQuery
@@ -283,13 +313,193 @@ class WikiDataCache extends BaseCache
      * Summary of parseEntity
      * @param array<mixed> $data
      * @param string $lang
-     * @return Entity
+     * @return Entity|array<mixed>|null
      */
     public static function parseEntity($data, $lang = 'en')
     {
-        $entity = new Entity($data, $lang);
+        if (empty($data['properties']) || empty($data['properties']['P31'])) {
+            return null;
+        }
+        $instanceOf = $data['properties']['P31'];
+        if (empty($instanceOf['values'])) {
+            return null;
+        }
+        $instanceType = null;
+        foreach ($instanceOf['values'] as $value) {
+            if (array_key_exists($value['id'], self::$instanceTypes)) {
+                $instanceType = self::$instanceTypes[$value['id']];
+                break;
+            }
+        }
+        if (empty($instanceType)) {
+            return null;
+        }
+        $data['type'] = $instanceType;
+        // @todo parse author, book, series etc.
+        $result = match ($instanceType) {
+            'author' => self::parseAuthor($data),
+            'book' => self::parseBook($data),
+            'series' => self::parseSeries($data),
+            'publisher' => self::parsePublisher($data),
+            default => throw new Exception('Unknown instance type ' . $instanceType),
+        };
+        return $result;
+        //$entity = new Entity($data, $lang);
         // @todo this generates warnings for missing prop, propertyLabel, qualifier etc.
-        $entity->parseProperties($data['properties'] ?? []);
-        return $entity;
+        //$entity->parseProperties($data['properties'] ?? []);
+        //return $entity;
+    }
+
+    /**
+     * Summary of parseAuthor
+     * @param array<mixed> $data
+     * @return array<mixed>|null
+     */
+    public static function parseAuthor($data)
+    {
+        $author = $data;
+        // @todo parse author, book, series etc.
+        unset($author['properties']);
+        return $author;
+    }
+
+    /**
+     * Summary of parseBook
+     * @param array<mixed> $data
+     * @return array<mixed>|null
+     */
+    public static function parseBook($data)
+    {
+        $book = $data;
+        // P50: author
+        $book['author'] = null;
+        if (!empty($book['properties']['P50'])) {
+            $book['author'] = $book['properties']['P50']['values'] ?? [];
+        }
+        // P407: language of work or name
+        $book['language'] = null;
+        if (!empty($book['properties']['P407']) && !empty($book['properties']['P407']['values'])) {
+            $book['language'] = $book['properties']['P407']['values'][0]['label'] ?? '';
+        }
+        // P577: publication date
+        $book['published'] = null;
+        if (!empty($book['properties']['P577'])) {
+            $dates = $book['properties']['P577']['values'];
+            if (!empty($dates)) {
+                $book['published'] = explode('T', $dates[0]['label'] ?? '')[0];
+            }
+        }
+        // P1476: title
+        //if (!empty($book['properties']['P1476'])) {
+        //    $book['title'] = $book['properties']['P1476']['values'] ?? [];
+        //}
+        // P179: part of the series
+        $book['series'] = null;
+        if (!empty($book['properties']['P179'])) {
+            $book['series'] = $book['properties']['P179']['values'] ?? [];
+            foreach ($book['series'] as $id => $series) {
+                $series['qualifiers'] ??= [];
+                foreach ($series['qualifiers'] as $qualifier) {
+                    if ($qualifier['id'] == 'P1545') {
+                        $book['series'][$id]['index'] = $qualifier['value'];
+                    }
+                }
+                unset($book['series'][$id]['qualifiers']);
+            }
+        }
+        // P123: publisher
+        $book['publisher'] = null;
+        if (!empty($book['properties']['P123']) && !empty($book['properties']['P123']['values'])) {
+            $book['publisher'] = $book['properties']['P123']['values'][0]['label'] ?? '';
+        }
+        // P18: image
+        $book['cover'] = null;
+        if (!empty($book['properties']['P18']) && !empty($book['properties']['P18']['values'])) {
+            $book['cover'] = $book['properties']['P18']['values'][0]['label'] ?? '';
+        }
+        // P136: genre
+        $book['genre'] = null;
+        if (!empty($book['properties']['P136'])) {
+            $book['genre'] = $book['properties']['P136']['values'] ?? [];
+        }
+        // P7937: form of creative work
+        $book['format'] = null;
+        if (!empty($book['properties']['P7937']) && !empty($book['properties']['P7937']['values'])) {
+            $book['format'] = $book['properties']['P7937']['values'][0]['label'] ?? '';
+        }
+        $book['identifiers'] = [];
+        $book['identifiers']['wd'] = $book['id'];
+        if (!empty($book['wiki_url'])) {
+            $book['identifiers']['url'] = $book['wiki_url'];
+        }
+        $book['identifiers'] = self::addBookIdentifiers($book['identifiers'], $book['properties']);
+        // @todo other properties
+
+        unset($book['properties']);
+        return $book;
+    }
+
+    /**
+     * Summary of addBookIdentifiers
+     * @param array<string, mixed> $identifiers
+     * @param array<string, mixed> $properties
+     * @return array<string, mixed>
+     */
+    public static function addBookIdentifiers($identifiers, $properties)
+    {
+        $todo = [
+            'isbn' => [
+                'P212',  // P212: ISBN-13
+                'P957',  // P957: ISBN-10
+            ],
+            'goodreads' => [
+                'P8383',  // P8383: Goodreads work ID
+                'P2969',  // P2969: Goodreads version\/edition ID
+            ],
+            'olid' => [
+                'P648',  // P648: Open Library ID
+            ],
+            'ltid' => [
+                'P1085',  // P1085: LibraryThing work ID
+            ],
+            'isfdb' => [
+                'P1274',  // P1274: ISFDB title ID
+            ],
+        ];
+        foreach ($todo as $name => $pids) {
+            foreach ($pids as $pid) {
+                if (!empty($properties[$pid]) && !empty($properties[$pid]['values'])) {
+                    $identifiers[$name] = $properties[$pid]['values'][0]['id'] ?? '';
+                    break;
+                }
+            }
+        }
+        return $identifiers;
+    }
+
+    /**
+     * Summary of parseSeries
+     * @param array<mixed> $data
+     * @return array<mixed>|null
+     */
+    public static function parseSeries($data)
+    {
+        $series = $data;
+        // @todo parse author, book, series etc.
+        unset($series['properties']);
+        return $series;
+    }
+
+    /**
+     * Summary of parsePublisher
+     * @param array<mixed> $data
+     * @return array<mixed>|null
+     */
+    public static function parsePublisher($data)
+    {
+        $publisher = $data;
+        // @todo parse author, book, series etc.
+        unset($publisher['properties']);
+        return $publisher;
     }
 }