+++ tests/GoogleBooksImportTest.php
@@ -12,6 +12,20 @@ use Marsender\EPubLoader\Import\JsonImport;
 
 class GoogleBooksImportTest extends BaseTestCase
 {
+    public function testJsonImportVolume(): void
+    {
+        $dbPath = dirname(__DIR__) . '/cache/google';
+        $dbFile = $dbPath . '/metadata.db';
+        $import = new JsonImport($dbFile, true);
+
+        $jsonFile = $dbPath . '/volumes/_ogXogEACAAJ.en.json';
+        [$message, $errors] = $import->loadFromJsonFile($dbPath, $jsonFile);
+
+        $expected = '/cache/google/volumes/_ogXogEACAAJ.en.json - 1 files OK - 0 files Error';
+        $this->assertStringContainsString($expected, $message);
+        $this->assertCount(0, $errors);
+    }
+
     public function testJsonImportFile(): void
     {
         $dbPath = dirname(__DIR__) . '/cache/google';
@@ -37,20 +51,6 @@ class GoogleBooksImportTest extends BaseTestCase
         [$message, $errors] = $import->loadFromPath($dbPath, $jsonPath);
 
         $expected = '/cache/google/titles/Émile Zola.La curée.fr.json - 10 files OK - 0 files Error';
-        $this->assertStringContainsString($expected, $message);
-        $this->assertCount(0, $errors);
-    }
-
-    public function testJsonImportVolume(): void
-    {
-        $dbPath = dirname(__DIR__) . '/cache/google';
-        $dbFile = $dbPath . '/metadata.db';
-        $import = new JsonImport($dbFile, true);
-
-        $jsonFile = $dbPath . '/volumes/_ogXogEACAAJ.en.json';
-        [$message, $errors] = $import->loadFromJsonFile($dbPath, $jsonFile);
-
-        $expected = '/cache/google/volumes/_ogXogEACAAJ.en.json - 1 files OK - 0 files Error';
         $this->assertStringContainsString($expected, $message);
         $this->assertCount(0, $errors);
     }