+++ tests/WikiDataImportTest.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * Epub loader application test
+ *
+ * @license    GPL v2 or later (http://www.gnu.org/licenses/gpl.html)
+ * @author     mikespub
+ */
+
+namespace Marsender\EPubLoader\Tests;
+
+use Marsender\EPubLoader\Import\JsonImport;
+
+class WikiDataImportTest extends BaseTestCase
+{
+    public function testJsonImportFile(): void
+    {
+        $dbPath = dirname(__DIR__) . '/cache/wikidata';
+        $dbFile = $dbPath . '/metadata.db';
+        $import = new JsonImport($dbFile, true);
+
+        $jsonFile = $dbPath . '/entities/Q223131.en.json';
+        [$message, $errors] = $import->loadFromJsonFile($dbPath, $jsonFile);
+
+        $expected = '/cache/wikidata/entities/Q223131.en.json - 1 files OK - 0 files Error';
+        $this->assertStringContainsString($expected, $message);
+        $this->assertCount(0, $errors);
+    }
+
+    public function testJsonImportPath(): void
+    {
+        $dbPath = dirname(__DIR__) . '/cache/wikidata';
+        $dbFile = $dbPath . '/metadata.db';
+        $import = new JsonImport($dbFile, true);
+
+        $jsonPath = 'entities';
+        [$message, $errors] = $import->loadFromPath($dbPath, $jsonPath);
+
+        $expected = '/cache/wikidata/entities/Q223131.en.json - 1 files OK - 0 files Error';
+        $this->assertStringContainsString($expected, $message);
+        $this->assertCount(0, $errors);
+    }
+}